BACKEND/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
sqlmodel>=0.0.8
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
email-validator>=1.3.1